Fixed power meter not being affected by angy

+++ Assets/Scripts/Ball/Shooter.cs
@@ -56,7 +56,8 @@ public class Shooter : MonoBehaviour{
     // Ball Spin
     private Vector3 spinDirection = new Vector3(0,0,0);
 
-
+    // \/ \/ \/ REMOVE, ONLY FOR TESTING \/ \/ \/
+    // /\ /\ /\ REMOVE, ONLY FOR TESTING /\ /\ /\
 
     public void SetBallFormActive(bool state)
     {
@@ -219,7 +220,7 @@ public class Shooter : MonoBehaviour{
         
     }
     private IEnumerator CalculateShootForce(){
-        float currentAngy = Mathf.Lerp(1,0.1f,GameManager.CurrentTurnPlayer.Angy / GameConfig.Instance.AngyValues.MaxAngy);
+        float currentAngy = Mathf.Lerp(1,0.1f, (float) GameManager.CurrentTurnPlayer.Angy / GameConfig.Instance.AngyValues.MaxAngy);
         forcePercent = 0;
         yield return null;
         while (!rewiredPlayer.GetButtonDown("Confirm") && forcePercent >= 0){
